requirements.txt
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-consumption==10.0.0
azure-mgmt-core==1.5.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.2.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.12.2
six==1.17.0
tabulate==0.9.0
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.3.8
